+++ README.md
@@ -6,7 +6,9 @@ Generic storage interface `Storer` for declarative resources.
 
 Generic storage is a public key-value storage HTTP server.
 
-Generic storage is intended for projects that are **in development**. This repository provides a minimal way (requiring few steps) of persisting arbitrary, unknown data online.
+Generic storage is intended for projects that are **in development**. This
+repository provides a minimal way (requiring few steps) of persisting arbitrary,
+unknown data online.
 
 ## Usage
 
@@ -16,7 +18,8 @@ _TODO(@EthanThatOneKid): Describe intended API usage by providing an example_.
 
 ### Deno Deploy
 
-_TODO(@EthanThatOneKid): Document the process of spinning up a simple, generic storage system on [Deno Deploy](https://deno.com/deploy/docs/deployments)._
+_TODO(@EthanThatOneKid): Document the process of spinning up a simple, generic
+storage system on [Deno Deploy](https://deno.com/deploy/docs/deployments)._
 
 ## Development
 
@@ -41,8 +44,7 @@ deno lint
 ### Running locally
 
 ```bash
-# TODO(@EthanThatOneKid): Create an example HTTP server that extends the given StorageInterface and example resource.
-# TODO(@EthanThatOneKid): Generate declarative resource via Protobuf hosted online.
+deno run --allow-net demos/default/main.ts
 ```
 
 ---
