+++ http/mod.ts
@@ -1,1 +1,2 @@
 export * from "./handler.ts";
+export * from "./server.ts";
