--- server/httpserver/mod.ts
@@ -1,1 +1,0 @@
-export * from "./server.ts";
