--- storer/mod.ts
@@ -1,1 +1,0 @@
-export * from "./storer.ts";
