--- storer/web_storer/mod.ts
@@ -1,1 +1,0 @@
-export { WebStorer } from "./storer.ts";
